refactor: move file_cleanup closer to unparser
